reverted to 16 bit precision

+++ model.py
@@ -10,13 +10,14 @@ import logging
 import os
 import torch.nn.functional as F
 
+
 ### Main model ###
 def load_model(model_name):
     try:
         model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=torch.float32,  # Use float16 for memory efficiency
-            device_map="auto",  # Automatically handle device placement
+            torch_dtype=torch.float16,
+            device_map="auto",
         )
         return model
     except Exception as e:
@@ -41,16 +42,19 @@ def load_model_and_tokenizer(model_name):
     tokenizer = load_tokenizer(model_name)
     return model, tokenizer
 
+
 # TODO: Add a class for speculative sampling model and a class for a chain of thought model
 ### Chain of Thought Model ###
 
 ### Entailment Model ###
+
 
 class BaseEntailment:
     """Base class for entailment models."""
 
     def save_prediction_cache(self):
         pass
+
 
 class EntailmentDeberta(BaseEntailment):
     """Entailment model using Deberta-v2-xlarge-mnli."""
@@ -72,4 +76,3 @@ class EntailmentDeberta(BaseEntailment):
             "neutral": probs[1].item(),
             "entailment": probs[2].item(),
         }
-
